clients_example/requirements.txt
termcolor==2.1
requests==2.30
sseclient-py==1.7